ongoing fine tuning

+++ nauiscaa_application/nausicaa_vs/Calibration/fbo.cpp
@@ -36,7 +36,7 @@ void FBO::Create(int w_,int h_)
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
 	glTexEnvi(GL_TEXTURE_ENV,GL_TEXTURE_ENV_MODE,GL_DECAL);
-	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB8, w, h, 0, GL_RGB, GL_UNSIGNED_BYTE, NULL);
+	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB32F, w, h, 0, GL_RGB, GL_UNSIGNED_BYTE, NULL);
 	
     glGenFramebuffersEXT(1, &this->id_fbo);
     glBindFramebufferEXT(GL_FRAMEBUFFER_EXT, this->id_fbo);
